requirements.txt
discord.py[voice]
asyncio
